refactor: get SuperNodes repo setup
