--- SuperNodes/src/SuperNode/SuperNodesRepo.cs
@@ -1,6 +1,0 @@
-namespace SuperNodes.SuperNode;
-
-/// <summary>
-/// Handles logic for generating SuperNodes.
-/// </summary>
-public class SuperNodesRepo { }
